Development1

+++ src/main/java/com/ispp/thorneo/service/impl/PromotionServiceImpl.java
@@ -1,19 +1,22 @@
 package com.ispp.thorneo.service.impl;
 
+import com.ispp.thorneo.domain.User;
 import com.ispp.thorneo.service.PromotionService;
 import com.ispp.thorneo.domain.Promotion;
 import com.ispp.thorneo.repository.PromotionRepository;
 import com.ispp.thorneo.repository.search.PromotionSearchRepository;
+import com.ispp.thorneo.service.UserService;
+import com.ispp.thorneo.web.rest.errors.BadRequestAlertException;
+import io.jsonwebtoken.lang.Assert;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import java.util.List;
 import java.util.Optional;
-import java.util.stream.Collectors;
-import java.util.stream.StreamSupport;
 
 import static org.elasticsearch.index.query.QueryBuilders.*;
 
@@ -30,9 +33,12 @@ public class PromotionServiceImpl implements PromotionService {
 
     private final PromotionSearchRepository promotionSearchRepository;
 
-    public PromotionServiceImpl(PromotionRepository promotionRepository, PromotionSearchRepository promotionSearchRepository) {
+    private final UserService userService;
+
+    public PromotionServiceImpl(PromotionRepository promotionRepository, PromotionSearchRepository promotionSearchRepository, UserService userService) {
         this.promotionRepository = promotionRepository;
         this.promotionSearchRepository = promotionSearchRepository;
+        this.userService = userService;
     }
 
     /**
@@ -52,13 +58,14 @@ public class PromotionServiceImpl implements PromotionService {
     /**
      * Get all the promotions.
      *
+     * @param pageable the pagination information
      * @return the list of entities
      */
     @Override
     @Transactional(readOnly = true)
-    public List<Promotion> findAll() {
+    public Page<Promotion> findAll(Pageable pageable) {
         log.debug("Request to get all Promotions");
-        return promotionRepository.findAll();
+        return promotionRepository.findAll(pageable);
     }
 
 
@@ -82,6 +89,15 @@ public class PromotionServiceImpl implements PromotionService {
      */
     @Override
     public void delete(Long id) {
+        User user = userService.getUserWithAuthorities().get();
+        Assert.notNull(user, "User is null");
+
+        Promotion promotion = promotionRepository.getOne(id);
+
+        if (promotion.getUser().getId() != user.getId()) {
+            throw new BadRequestAlertException("invalid user", "promotion", "notCreator");
+        }
+
         log.debug("Request to delete Promotion : {}", id);        promotionRepository.deleteById(id);
         promotionSearchRepository.deleteById(id);
     }
@@ -90,14 +106,32 @@ public class PromotionServiceImpl implements PromotionService {
      * Search for the promotion corresponding to the query.
      *
      * @param query the query of the search
+     * @param pageable the pagination information
      * @return the list of entities
      */
     @Override
     @Transactional(readOnly = true)
-    public List<Promotion> search(String query) {
-        log.debug("Request to search Promotions for query {}", query);
-        return StreamSupport
-            .stream(promotionSearchRepository.search(queryStringQuery(query)).spliterator(), false)
-            .collect(Collectors.toList());
+    public Page<Promotion> search(String query, Pageable pageable) {
+        log.debug("Request to search for a page of Promotions for query {}", query);
+        return promotionSearchRepository.search(queryStringQuery(query), pageable);    }
+
+    @Override
+    public Promotion savePromotion(Promotion promotion) {
+        Promotion result;
+
+        Assert.notNull(promotion, "Promotion is null");
+
+        User user = userService.getUserWithAuthorities().get();
+        Assert.notNull(user, "User is null");
+
+        if (promotion.getUser() != null && promotion.getUser().getId() != user.getId()) {
+            throw new BadRequestAlertException("invalid user", "promotion", "notCreator");
+        }
+
+        promotion.setUser(user);
+
+        result = save(promotion);
+
+        return result;
     }
 }
